errors have been fixed
