+++ app/src/main/java/com/practicum/playlistmaker/favorites/data/FavoritesRepositoryImpl.kt
@@ -1,7 +1,8 @@
 package com.practicum.playlistmaker.favorites.data
 
-import com.practicum.playlistmaker.favorites.data.db.FavoriteTrackEntity
 import com.practicum.playlistmaker.favorites.data.db.FavoritesDao
+import com.practicum.playlistmaker.favorites.data.mapper.toDomainTrack
+import com.practicum.playlistmaker.favorites.data.mapper.toEntity
 import com.practicum.playlistmaker.favorites.domain.api.FavoritesRepository
 import com.practicum.playlistmaker.player.domain.models.Track
 import kotlinx.coroutines.flow.Flow
@@ -32,38 +33,6 @@ class FavoritesRepositoryImpl(
 
     override suspend fun isFavorite(trackId: Long): Boolean {
         return favoritesDao.isFavorite(trackId)
-    }
-
-    private fun Track.toEntity(): FavoriteTrackEntity {
-        return FavoriteTrackEntity(
-            trackId = trackId,
-            artworkUrl100 = artworkUrl100,
-            trackName = trackName,
-            artistName = artistName,
-            collectionName = collectionName,
-            releaseDate = releaseDate,
-            primaryGenreName = primaryGenreName,
-            country = country,
-            trackTimeMillis = trackTimeMillis,
-            previewUrl = previewUrl
-        )
-    }
-
-    private fun FavoriteTrackEntity.toDomainTrack(): Track {
-        return Track(
-            trackId = trackId,
-            trackName = trackName,
-            artistName = artistName,
-            trackTimeMillis = trackTimeMillis,
-            artworkUrl100 = artworkUrl100,
-            collectionName = collectionName ?: "",
-            releaseDate = releaseDate ?: "",
-            primaryGenreName = primaryGenreName ?: "",
-            country = country ?: "",
-            previewUrl = previewUrl ?: ""
-        ).apply {
-            isFavorite = true
-        }
     }
 }
