+++ app/src/main/java/com/practicum/playlistmaker/main/viewmodel/FavoritesUiState.kt
@@ -1,0 +1,9 @@
+package com.practicum.playlistmaker.main.viewmodel
+
+import com.practicum.playlistmaker.player.domain.models.Track
+
+data class FavoritesUiState(
+    val tracks: List<Track> = emptyList(),
+    val isEmpty: Boolean = true
+)
+
