+++ app/src/main/java/com/practicum/playlistmaker/main/viewmodel/SearchViewModel.kt
@@ -4,7 +4,6 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.practicum.playlistmaker.favorites.domain.api.FavoritesInteractor
 import com.practicum.playlistmaker.player.domain.api.SearchHistoryRepository
 import com.practicum.playlistmaker.player.domain.api.TrackInteractor
 import com.practicum.playlistmaker.player.domain.models.Track
@@ -25,8 +24,7 @@ data class SearchUiState(
 
 class SearchViewModel(
     private val trackInteractor: TrackInteractor,
-    private val searchHistoryRepository: SearchHistoryRepository,
-    private val favoritesInteractor: FavoritesInteractor
+    private val searchHistoryRepository: SearchHistoryRepository
 ) : ViewModel() {
 
     companion object {
@@ -98,11 +96,7 @@ class SearchViewModel(
 
     fun fetchHistory() {
         viewModelScope.launch {
-            val fetchedHistory = searchHistoryRepository.getHistory().toMutableList()
-            val favoriteIds = favoritesInteractor.getFavoriteTrackIds()
-            fetchedHistory.forEach { track ->
-                track.isFavorite = favoriteIds.contains(track.trackId)
-            }
+            val fetchedHistory = searchHistoryRepository.getHistory()
             val current = _state.value ?: SearchUiState()
             val isHistoryVisible = fetchedHistory.isNotEmpty() && current.searchText.isBlank()
             _state.value = current.copy(
