+++ app/src/main/res/drawable-night/circle_large.xml
@@ -2,7 +2,7 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="oval">
     <size
-        android:width="100dp"
-        android:height="100dp"/>
+        android:width="80dp"
+        android:height="80dp"/>
     <solid android:color="@color/white"/>
 </shape>
