+++ app/src/main/res/drawable/baseline_play_arrow_24.xml
@@ -1,5 +1,5 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="100dp" android:tint="@color/white" android:viewportHeight="100" android:viewportWidth="100" android:width="100dp">
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="60dp" android:tint="@color/white" android:viewportHeight="100" android:viewportWidth="100" android:width="60dp">
       
-    <path android:fillColor="@color/white" android:pathData="M32,20v56l44,-28z"/>
-
+    <path android:fillColor="@color/white" android:pathData="M30,15 L30,85 L75,50 Z"/>
+    
 </vector>
