+++ app/src/main/res/drawable/bottom_nav_item_background.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_pressed="true" android:drawable="@color/background_color"/> <!-- Прозрачный при нажатии -->
+    <item android:drawable="@color/background_color"/> <!-- Прозрачный фон по умолчанию -->
+</selector>
