+++ app/src/main/res/drawable/circle_large.xml
@@ -2,8 +2,8 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="oval">
     <size
-        android:width="100dp"
-        android:height="100dp"/>
+        android:width="80dp"
+        android:height="80dp"/>
     <solid android:color="@color/black"/>
 </shape>
