+++ app/src/main/res/layout/fragment_track_details.xml
@@ -90,8 +90,8 @@ android:layout_marginHorizontal="@dimen/twenty_four" >
 
 <ImageView
     android:id="@+id/playButton"
-    android:layout_width="100dp"
-    android:layout_height="100dp"
+    android:layout_width="80dp"
+    android:layout_height="80dp"
     android:src="@drawable/baseline_play_arrow_24"
     android:background="@drawable/circle_large"
     android:contentDescription="@string/play"
